move android stuff to view class
