+++ diesel/src/main/java/com/nuvoex/diesel/core/LoginPresenter.java
@@ -1,19 +1,8 @@
 package com.nuvoex.diesel.core;
 
-import android.content.ComponentName;
-import android.content.Context;
-import android.content.Intent;
-import android.content.pm.ActivityInfo;
-import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager;
-import android.content.pm.ResolveInfo;
-import android.net.Uri;
 import android.support.annotation.NonNull;
-import android.util.Log;
 
 import com.nuvoex.diesel.model.LoginResponse;
-
-import java.util.List;
 
 
 public class LoginPresenter implements LoginContract.Presenter {
@@ -91,35 +80,6 @@ public class LoginPresenter implements LoginContract.Presenter {
         return USERNAME_LENGTH;
     }
 
-    @Override
-    public void redirectToPlayStore() {
-        String playStoreID = mLoginView.getApplicationContext().getPackageName();
-
-        Intent rateIntent = new Intent(Intent.ACTION_VIEW, Uri.parse("market://details?id=" + playStoreID));
-        boolean isPlayStoreAppPresent = false;
-        // Need to directly open play store app only as more apps may have handled market:// Urls
-        final List<ResolveInfo> otherApps = mLoginView.getApplicationContext().getPackageManager().queryIntentActivities(rateIntent, 0);
-        for (ResolveInfo otherApp : otherApps)
-            if (otherApp.activityInfo.applicationInfo.packageName.equals("com.android.vending")) {
-
-                ActivityInfo otherAppActivity = otherApp.activityInfo;
-                ComponentName componentName = new ComponentName(
-                        otherAppActivity.applicationInfo.packageName,
-                        otherAppActivity.name
-                );
-                rateIntent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_MULTIPLE_TASK | Intent.FLAG_ACTIVITY_NO_HISTORY);
-                rateIntent.setComponent(componentName);
-                isPlayStoreAppPresent = true;
-                break;
-            }
-        if (!isPlayStoreAppPresent) {
-            rateIntent = new Intent(Intent.ACTION_VIEW, Uri.parse("https://play.google.com/store/apps/details?id=" + playStoreID));
-        }
-        mLoginView.getApplicationContext().startActivity(rateIntent);
-
-
-    }
-
     private Repository.LoginCallback mLoginCallback = new Repository.LoginCallback() {
         @Override
         public void successful(LoginResponse loginResponse) {
@@ -147,20 +107,8 @@ public class LoginPresenter implements LoginContract.Presenter {
 
         @Override
         public boolean isLatestApk(int latestVersionCode) {
+            return mLoginView.isLatestApk(latestVersionCode);
 
-            Context context = mLoginView.getApplicationContext();
-            int version;
-            try {
-                PackageInfo pInfo = context.getPackageManager().getPackageInfo(context.getPackageName(), 0);
-                version = pInfo.versionCode;
-            } catch (PackageManager.NameNotFoundException e) {
-                e.printStackTrace();
-                return true;
-            }
-            if (latestVersionCode > version) {
-                return false;
-            }
-            return true;
         }
 
         @Override
